+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
-## 0.7.0
+## 0.8.0
 ### Changed
 - Added enum support
-  
+
+## 0.7.0
+### Added
+- Optional arguments can now be skipped being passed into the TypedDict class constructor.
+- Generated code is automatically formatted with `black`.
+
+### Changed
+- The `astor` library has been replaced with `astunparse`.
+
 ## 0.6.0
 ### Changed
 - Make library compatible with Python 3.8